Change API to support multiple servers
This commit adds `*-to` versions of all of the various `add-handler`
functions in the service protocol.  These new functions take an
extra argument, `server-id`, and will be used to allow callers to
configure multiple servers and register handlers on a specific server.
The corresponding functions in `core` have also been modified to
require this extra `server-id` param.

None of the implementation for multiple servers is actually in
place yet; this commit simply changes the API and gets the tests
passing with the new API.